SWIP-1109 format empty db activity page (#255)
* SWIP-1109 Empty h1 tag on db page with no entries

This change is for all users and includes css overrides to ensure there is no white space with the empty h1. Could not remove the h1 entirely as mustache template is used on other pages that require the h1.

* Move breadcrumb logic from govuk to child theme

* SWIP-1109 simplify empty heading logic

And apply to the database page in general, as per designs

* SWIP-1109 Hide groups dropdown for non staff users

Hiding on the view database page. Refactored primary_nav to make the is_staff_like function a helper method.

* SWIP-1109 Change button name

Language string file will need importing into each environment. We would ideally automate.

* SWIP-1109 Remove elements that are not needed

For the empty state of  database activity

* SWIP-1109 Add style overrides via scss

Use Moodle's scss compiler

* SWIP-1109 Lib file comments

* SWIP-1109 Button text change

* SWIP-1109 Bump up theme version