SWIP-984 Replace file sync mechanism with NFS file share (#260)
The existing file synch copies files from a shared network folder to the moodledata folder. This solution creates a new NFS file share that maps to the moodledata folder directly. There is no need to copy files around and also provides us with the ability to use multiple Moodle instances and to backup the moodledata folder.